fixed calibration bug's & cosmetic changes

+++ ros/src/sensing/fusion/packages/calibration_camera_lidar/nodes/calibration_test/common_2d_calib.cpp
@@ -30,9 +30,9 @@
 
 #include "common_2d_calib.h"
 
-CvMat *v_g2l = cvCreateMat (1, 3, CV_32FC1);
-CvMat *v_g2c = cvCreateMat (1, 3, CV_32FC1);
-CvMat *v_rotation = cvCreateMat (1, 3, CV_32FC1);
+CvMat *v_g2l = cvCreateMat (1, 3, CV_64FC1);
+CvMat *v_g2c = cvCreateMat (1, 3, CV_64FC1);
+CvMat *v_rotation = cvCreateMat (1, 3, CV_64FC1);
 CvMat *m_rotation = cvCreateMat (3, 3, CV_64FC1);
 CvMat *m_intrinsic = cvCreateMat (3, 3, CV_64FC1);
 CvMat *m_dist = cvCreateMat (1, 4, CV_64FC1);
